@@ -1,9 +1,10 @@
 {
   "name": "table-to-excel-react",
-  "version": "1.1.1",
+  "version": "1.2.0",
   "description": "Export table data to Excel file using ReactJS",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
+  "type": "module",
   "private": false,
   "scripts": {
     "build": "tsc",
@@ -13,11 +14,11 @@
     "test": "jest",
     "test:watch": "jest --watch",
     "test:coverage": "jest --coverage",
-    "lint": "eslint src --ext .ts,.tsx",
-    "lint:fix": "eslint src --ext .ts,.tsx --fix",
+    "lint": "eslint --config eslint.config.js src --ext .ts,.tsx",
+    "lint:fix": "eslint --config eslint.config.js src --ext .ts,.tsx --fix",
     "format": "prettier --write \"src/**/*.{ts,tsx}\"",
     "format:all": "prettier --write \"src/**/*.{ts,tsx,js,jsx,json,md}\" && npm run lint:fix",
-    "format:script": "node format-code.js",
+    "format:script": "node format-code.cjs",
     "prepare": "npm run build",
     "prepublishOnly": "npm run format:script",
     "preversion": "npm run lint",
@@ -25,34 +26,35 @@
     "postversion": "git push && git push --tags"
   },
   "peerDependencies": {
-    "react": "^17.0.0 || ^18.0.0 || ^19.0.0"
+    "react": "^18.0.0 || ^19.0.0"
   },
   "devDependencies": {
-    "@babel/preset-typescript": "^7.17.12",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^14.0.0",
-    "@types/jest": "^29.5.0",
-    "@types/node": "^17.0.39",
-    "@types/react": "^18.0.10",
-    "@typescript-eslint/eslint-plugin": "^5.59.0",
-    "@typescript-eslint/parser": "^5.59.0",
-    "eslint": "^8.38.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-react": "^7.32.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "jest": "^29.5.0",
-    "jest-environment-jsdom": "^29.5.0",
-    "prettier": "^2.8.8",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
-    "rimraf": "^5.0.1",
-    "ts-jest": "^29.1.0",
-    "typescript": "^4.7.2"
+    "@babel/preset-typescript": "^7.26.0",
+    "@eslint/js": "^9.22.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.2.0",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.13.10",
+    "@types/react": "^19.0.10",
+    "@typescript-eslint/eslint-plugin": "^8.26.0",
+    "@typescript-eslint/parser": "^8.26.0",
+    "eslint": "^9.22.0",
+    "eslint-config-prettier": "^10.1.1",
+    "eslint-plugin-prettier": "^5.2.3",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "prettier": "^3.5.3",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "rimraf": "^6.0.1",
+    "ts-jest": "^29.2.6",
+    "typescript": "^5.8.2",
+    "typescript-eslint": "^8.26.0"
   },
   "dependencies": {
-    "xlsx": "^0.18.5",
-    "xlsx-style": "^0.8.13"
+    "exceljs": "^4.4.0"
   },
   "repository": {
     "type": "git",
